+++ sknetwork/data/loading.py
@@ -6,14 +6,20 @@ Created on November 15, 2019
 This module's code is freely adapted from TorchKGE's torchkge.data.DataLoader.py code.
 """
 
+import tarfile
 import shutil
 import zipfile
-from os import environ, makedirs, remove
+from os import environ, makedirs, remove, listdir
 from os.path import exists, expanduser, join
 from typing import Optional
+from urllib.error import HTTPError
 from urllib.request import urlretrieve
 
-from sknetwork.data.parsing import parse_tsv, parse_labels, parse_hierarchical_labels
+import numpy as np
+
+from sknetwork.basics import is_bipartite
+from sknetwork.data.parsing import parse_tsv, parse_labels, parse_hierarchical_labels, parse_header, parse_metadata
+from sknetwork.utils import Bunch
 
 
 def get_data_home(data_home: Optional[str] = None):
@@ -47,75 +53,127 @@ def clear_data_home(data_home: Optional[str] = None):
     shutil.rmtree(data_home)
 
 
-def load_vital_wikipedia(data_home: Optional[str] = None, outputs: str = 'both', return_titles: bool = False,
-                         return_labels: bool = False, return_categories: bool = False, max_depth: int = 1,
-                         full_path: bool = True):
+def load_wikilinks_dataset(dataset_name: str, data_home: Optional[str] = None,
+                           max_depth: int = 1, full_path: bool = True):
     """
-    Loads the Vital Wikipedia dataset. See https://graphs.telecom-paristech.fr/
-    Data returned in this order (depending on the selection): adjacency, biadjacency, titles, stems, categories.
+    Loads a dataset from the `WikiLinks database
+    <https://graphs.telecom-paristech.fr/Home_page.html#wikilinks-section>`_.
 
     Parameters
     ----------
+    dataset_name: str
+        The name of the dataset (no capital letters, all spaces replaced with underscores)
     data_home: str
         The folder to be used for dataset storage
-    outputs: str
-        Defines what should be returned. Possible options are 'adjacency' for the adjacency of hyperlinks of the various
-        Wikipedia articles, 'biadjacency' for the count matrix of the stems or 'both' for both.
-    return_titles: bool
-        If ``True``, returns the titles of the articles
-    return_labels: bool
-        If ``True``, returns the titles of the articles and the stems of the summary
-    return_categories: bool
-        If ``True``, returns the categories of the articles
     max_depth: int
-        Denotes the maximum depth to use for the categories
+        Denotes the maximum depth to use for the categories (if relevant)
     full_path: bool
         Denotes if only the deepest label possible should be returned or if all super categories should
-        be considered (default)
+        be considered (if relevant)
 
     Returns
     -------
-    adjacency: sparse.csr_matrix, optional
-        The matrix of the hyperlinks between the articles
-    biadjacency: sparse.csr_matrix, optional
-        The matrix of the stem counts in each summary of article
-    titles: numpy.ndarray, optional
-        The array of the article titles
-    stems: numpy.ndarray, optional
-        The array of the stems
-    categories: numpy.ndarray, optional
-        The array of the categories
+    data: :class:`Bunch`
+        An object with some of the following attributes (depending on the dataset):
+
+         * `adjacency`: the adjacency matrix of the graph in CSR format
+         * `biadjacency`: the biadjacency matrix of the graph in CSR format
+         * `feature_names`: the array of the names for the features
+         * `names`: the titles of the articles
+         * `target_names`: the categories of the articles as specified with `max_depth` and `full_path`
+         * `target`: the index for `target_names`
+
     """
     if data_home is None:
         data_home = get_data_home()
-    data_path = data_home + '/vital_wikipedia'
+    data_path = data_home + '/' + dataset_name + '/'
     if not exists(data_path):
         makedirs(data_path, exist_ok=True)
-        urlretrieve("https://graphs.telecom-paristech.fr/datasets/vital_wikipedia.zip",
-                    data_home + '/vital_wikipedia.zip')
-        with zipfile.ZipFile(data_home + '/vital_wikipedia.zip', 'r') as zip_ref:
-            zip_ref.extractall(data_home)
-        remove(data_home + '/vital_wikipedia.zip')
+        try:
+            urlretrieve("https://graphs.telecom-paristech.fr/datasets/" + dataset_name + '.tar.gz',
+                        data_home + '/' + dataset_name + '.tar.gz')
+        except HTTPError:
+            raise ValueError('Invalid dataset ' + dataset_name)
+        with tarfile.open(data_home + '/' + dataset_name + '.tar.gz', 'r:gz') as tar_ref:
+            tar_ref.extractall(data_home)
+        remove(data_home + '/' + dataset_name + '.tar.gz')
 
-    if outputs == 'adjacency':
-        output = [parse_tsv(data_path + '/en-internal-links.txt', directed=True, reindex=False)]
-    elif outputs == 'biadjacency':
-        output = [parse_tsv(data_path + '/en-articles-stems.txt', bipartite=True, reindex=False)]
-    elif outputs == 'both':
-        output = [parse_tsv(data_path + '/en-internal-links.txt', directed=True, reindex=False),
-                  parse_tsv(data_path + '/en-articles-stems.txt', bipartite=True, reindex=False)]
-    else:
-        raise ValueError("Outputs must be 'adjacency', 'biadjacency' or 'both'.")
+    data = Bunch()
+    files = [file for file in listdir(data_path)]
 
-    if return_labels:
-        output.append(parse_labels(data_path + '/en-articles.txt'))
-        output.append(parse_labels(data_path + '/en-stems.txt'))
-    elif return_titles:
-        output.append(parse_labels(data_path + '/en-articles.txt'))
-    if return_categories:
-        output.append(parse_hierarchical_labels(data_path + '/en-categories.txt', max_depth, full_path=full_path))
+    if 'adjacency.txt' in files:
+        data.adjacency = parse_tsv(data_path + '/adjacency.txt', directed=True, reindex=False)
+    if 'biadjacency.txt' in files:
+        data.biadjacency = parse_tsv(data_path + '/biadjacency.txt', bipartite=True, reindex=False)
+    if 'names.txt' in files:
+        data.names = parse_labels(data_path + '/names.txt')
+    if 'feature_names.txt' in files:
+        data.feature_names = parse_labels(data_path + '/feature_names.txt')
+    if 'categories.txt' in files:
+        data.target_names = parse_hierarchical_labels(data_path + '/categories.txt', max_depth, full_path=full_path)
+        _, data.target = np.unique(data.target_names, return_inverse=True)
 
-    if len(output) == 1:
-        return output[0]
-    else:
-        return tuple(output)
+    return data
+
+
+def load_konect_dataset(dataset_name: str, data_home: Optional[str] = None):
+    """
+    Loads a dataset from the `Konect database
+    <http://konect.uni-koblenz.de>`_.
+
+    Parameters
+    ----------
+    dataset_name: str
+        The name of the dataset as specified in the download link (e.g. for the Actor movies dataset, the corresponding
+        name is ``'actor-movie'``).
+    data_home: str
+        The folder to be used for dataset storage
+
+    Returns
+    -------
+    data: :class:`Bunch`
+        An object with the following attributes:
+
+         * `adjacency` or `biadjacency`: the adjacency/biadjacency matrix for the dataset
+         * `meta`: a dictionary containing the metadata as specified by Konect
+         * any attribute described in an ent.* file
+
+    """
+    if data_home is None:
+        data_home = get_data_home()
+    data_path = data_home + '/' + dataset_name + '/'
+    if not exists(data_path):
+        makedirs(data_path, exist_ok=True)
+        try:
+            urlretrieve('http://konect.uni-koblenz.de/downloads/tsv/' + dataset_name + '.tar.bz2',
+                        data_home + '/' + dataset_name + '.tar.bz2')
+        except HTTPError:
+            raise ValueError('Invalid dataset ' + dataset_name)
+        with tarfile.open(data_home + '/' + dataset_name + '.tar.bz2', 'r:bz2') as tar_ref:
+            tar_ref.extractall(data_home)
+        remove(data_home + '/' + dataset_name + '.tar.bz2')
+
+    data = Bunch()
+    files = [file for file in listdir(data_path) if dataset_name in file]
+
+    matrix = [file for file in files if 'out.' in file]
+    if matrix:
+        file = matrix[0]
+        directed, bipartite, weighted = parse_header(data_path + file)
+        if bipartite:
+            data.biadjacency = parse_tsv(data_path + file, directed=directed, bipartite=bipartite, weighted=weighted)[0]
+        else:
+            data.adjacency = parse_tsv(data_path + file, directed=directed, bipartite=bipartite, weighted=weighted)[0]
+
+    metadata = [file for file in files if 'meta.' in file]
+    if metadata:
+        file = metadata[0]
+        data.meta = parse_metadata(data_path + file)
+
+    attributes = [file for file in files if 'ent.' + dataset_name in file]
+    if attributes:
+        for file in attributes:
+            attribute_name = file.split('.')[-1]
+            data[attribute_name] = parse_labels(data_path + file)
+
+    return data