+++ sknetwork/utils/base.py
@@ -11,10 +11,9 @@ class Algorithm:
     """
     Base class for all algorithms
     """
-
     def __repr__(self):
         # parameters not to display
-        arg_black_list = ['self', 'random_state']
+        arg_black_list = ['self', 'random_state', 'verbose']
         output = self.__class__.__name__ + '('
         signature = inspect.signature(self.__class__.__init__)
         arguments = [arg.name for arg in signature.parameters.values() if arg.name not in arg_black_list]