+++ sknetwork/utils/tests/test_verbose.py
@@ -1,0 +1,16 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""tests for verbose.py"""
+
+import unittest
+
+from sknetwork.utils.verbose import VerboseMixin
+
+
+# noinspection PyMissingOrEmptyDocstring
+class TestChecks(unittest.TestCase):
+
+    def test_prints(self):
+        verbose = VerboseMixin(verbose=True)
+        verbose.log.print('There are', 4, 'seasons in a year')
+        self.assertEqual(str(verbose.log), 'There are 4 seasons in a year\n')